hotfix: do not raise error when all coverages have no datetime_ranges
